feat: big-map: add instructions at map footer
