+++ scene/src/ui-classes/main-hud/big-map/big-map-view.tsx
@@ -53,6 +53,7 @@ import {
 import { fetchLiveEvents } from '../../../utils/fetch-live-events'
 import { EventFromApi } from '../../scene-info-card/SceneInfoCard.types'
 import { currentRealmProviderIsWorld } from '../../../service/realm-change'
+import { MapFooter } from './map-footer'
 
 export const FOV = (45 * 1.25 * Math.PI) / 180
 
@@ -478,142 +479,4 @@ export type MapPinParams = {
 }
 function MapPin({ sprite, size, position, isActive, onClick }: MapPinParams) {
   // TODO
-}
-
-export function MapFooter(): ReactElement {
-  const canvasScaleRatio = getCanvasScaleRatio()
-  return (
-    <UiEntity
-      uiTransform={{
-        positionType: 'absolute',
-        position: {
-          bottom: 0,
-          left: 0
-        },
-        width: '100%',
-        height: getViewportHeight() * 0.04,
-        flexDirection: 'row',
-        alignItems: 'center',
-        alignContent: 'center',
-        justifyContent: 'center'
-      }}
-      uiBackground={{
-        color: COLOR.DARK_OPACITY_5
-      }}
-    >
-      <MapStatusBar />
-      <UiEntity
-        uiTransform={{
-          flexDirection: 'row',
-          justifyContent: 'flex-end',
-          alignItems: 'flex-end'
-        }}
-      >
-        <Icon
-          icon={{ atlasName: 'icons', spriteName: 'LeftClick' }}
-          iconSize={canvasScaleRatio * 26}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: canvasScaleRatio * 13 }
-          }}
-        />
-        <UiEntity
-          uiText={{
-            value: 'Double click to displace map to target',
-            fontSize: canvasScaleRatio * 26
-          }}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: 0 }
-          }}
-        />
-      </UiEntity>
-      <UiEntity
-        uiTransform={{
-          flexDirection: 'row',
-          justifyContent: 'flex-end',
-          alignItems: 'flex-end'
-        }}
-      >
-        <Icon
-          icon={{ atlasName: 'icons', spriteName: 'LeftClick' }}
-          iconSize={canvasScaleRatio * 26}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: canvasScaleRatio * 13 }
-          }}
-        />
-        <UiEntity
-          uiText={{
-            value: 'Drag to displace map',
-            fontSize: canvasScaleRatio * 26
-          }}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: 0 }
-          }}
-        />
-      </UiEntity>
-      <UiEntity
-        uiTransform={{
-          flexDirection: 'row',
-          justifyContent: 'flex-end',
-          alignItems: 'flex-end'
-        }}
-      >
-        <Icon
-          icon={{ atlasName: 'icons', spriteName: 'Scroll' }}
-          iconSize={canvasScaleRatio * 26}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: canvasScaleRatio * 13 }
-          }}
-        />
-        <UiEntity
-          uiText={{
-            value: 'Mouse wheel to zoom in/out',
-            fontSize: canvasScaleRatio * 26
-          }}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: 0 }
-          }}
-        />
-      </UiEntity>
-      <UiEntity
-        uiTransform={{
-          flexDirection: 'row',
-          justifyContent: 'flex-end',
-          alignItems: 'flex-end'
-        }}
-      >
-        <Icon
-          icon={{ atlasName: 'icons', spriteName: 'UpArrow' }}
-          iconSize={canvasScaleRatio * 26}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: canvasScaleRatio * 13 }
-          }}
-        />
-        <UiEntity
-          uiText={{
-            value: 'Shift + Mouse move to rotate view',
-            fontSize: canvasScaleRatio * 26
-          }}
-          uiTransform={{
-            alignSelf: 'flex-start',
-            justifyContent: 'center',
-            position: { top: 0 }
-          }}
-        />
-      </UiEntity>
-    </UiEntity>
-  )
 }
